Random

+++ G3D.lib/source/System.cpp
@@ -1469,7 +1469,7 @@ public:
                 System::outOfMemoryCallback(REALBLOCK_SIZE(bytes), false);
             }
 #           ifdef G3D_DEBUG
-                debugPrintf("::malloc(%d) returned NULL\n", REALBLOCK_SIZE(bytes));
+            debugPrintf("::malloc(%d) returned NULL\n", (int)REALBLOCK_SIZE(bytes));
 #           endif
             debugAssertM(ptr != NULL, 
                          "::malloc returned NULL. Either the "
